update JUnit testing
