+++ data/TaskList.txt
@@ -1,4 +1,1 @@
-[T][ ] read book
-[T][X] return book
-[D][X] do assignment (by: sunday)
-[E][X] kpop (from: 12pm to: 2pm)
+[T][ ] sleep
