+++ src/test/java/duke/ParserTest.java
@@ -11,11 +11,4 @@ public class ParserTest {
         ExitTask tsk = new ExitTask();
         assertEquals(tsk, c);
     }
-
-    @Test
-    public void deleteTest() throws InvalidInputException {
-        Command c = Parser.parse("delete 3");
-        DeleteTask tsk = new DeleteTask(3);
-        assertEquals(tsk, c);
-    }
 }
