+++ src/test/java/duke/ToDoTest.java
@@ -11,15 +11,4 @@ public class ToDoTest {
         tsk.setCompleted();
         assertEquals("[X] ", tsk.getCheckbox());
     }
-
-    @Test
-    public void executeTest() {
-        Ui ui = new Ui();
-        Storage storage = new Storage("data/tasks.txt");
-        TaskList lst = new TaskList();
-
-        ToDo tsk = new ToDo("sleep");
-        tsk.execute(lst, ui, storage);
-        assertEquals(1, lst.size());
-    }
 }
